+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## 1.1.0
+- added action `shgi/GitPackages/updatePackage/success`
+- added action `shgi/GitPackages/updatePackage/error`
+- bugfix: update log did not save prevVersion
 
 ## 1.0.0